backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
requests
beautifulsoup4
googlesearch-python
google-search-results 
serpapi
sentence-transformers
numpy
scipy
pydantic
